dev-requirements.txt
docopt==0.6.2
Jinja2==2.9.6
MarkupSafe==1.0
pyinotify==0.9.6
PyYAML==3.12
